# cholesky_solve_backward: speed up using output_mask