--- a/test/test_linalg.py
+++ b/test/test_linalg.py
@@ -2637,6 +2637,27 @@ class TestLinalg(TestCase):
             self.assertEqual(len(w), 1)
             self.assertTrue("An output with one or more elements was resized" in str(w[-1].message))
 
+    @skipCUDAIfNoMagma
+    @skipCPUIfNoLapack
+    @dtypes(torch.float, torch.double)
+    def test_cholesky_solve_backward(self, device, dtype):
+        from torch.testing._internal.common_utils import random_hermitian_pd_matrix
+
+        # A different code path runs depending on if we ask for the grad of A.
+        for requires_A_grad in (False, True):
+            b_dims = (5, 2)
+            A_dims = (5,)
+            upper = False
+
+            b = torch.randn(*b_dims, dtype=dtype, device=device, requires_grad=True)
+            A = random_hermitian_pd_matrix(*A_dims, dtype=dtype, device=device)
+            A.requires_grad = requires_A_grad
+            L = torch.linalg.cholesky(A, upper=upper)
+            (torch.cholesky_solve(b, L, upper=upper).sum()).backward()
+
+            expected_grad = torch.linalg.solve(A, torch.ones_like(b))
+            self.assertEqual(b.grad, expected_grad)
+
     @skipCUDAIfNoMagmaAndNoCusolver
     @skipCPUIfNoLapack
     @dtypes(*floating_and_complex_types())