--- a/torch/csrc/autograd/FunctionsManual.h
+++ b/torch/csrc/autograd/FunctionsManual.h
@@ -743,7 +743,8 @@ std::tuple<Tensor, Tensor> cholesky_solve_backward(
     const Tensor& self,
     const Tensor& input2,
     const Tensor& result,
-    const bool upper);
+    const bool upper,
+    std::array<bool, 2> output_mask);
 Tensor cholesky_solve_jvp(
     const Tensor& X,
     const Tensor& U,