@lezcano 2023-11-16
Yep, I also thought about that nit with `upper=False`. Thank you for fixing it!